Proper dispoing while handling collection load and unload for vector first frontend test for retrival fix collection deregister

+++ IqraCore/Models/RAG/Retrieval/RAGRetrievalSourceModel.cs
@@ -6,6 +6,6 @@
         public string DocumentName { get; set; } = string.Empty;
         public string ChunkId { get; set; } = string.Empty;
         public string Content { get; set; } = string.Empty;
-        public double Score { get; set; }
+        public float Score { get; set; }
     }
 }
